gui copy/requirements.txt
streamlit>=1.28.0
aider-chat>=0.61.0
pylint>=3.0.0
bandit>=1.7.5
semgrep>=1.34.0
gitleaks
docker>=6.1.3
psutil>=5.9.0
sysstat>=12.0.0  # For iostat command
netifaces>=0.11.0  # For network interface info
gitpython>=3.1.40  # For advanced git operations
memory-profiler>=0.61.0  # For detailed memory analysis
py-cpuinfo>=9.0.0  # For CPU information
